Split test_category file into smaller test files
